+++ js/ui/scheduler/recurrence.js
@@ -6,7 +6,6 @@ import { inArray } from '../../core/utils/array';
 import { RRule, RRuleSet } from 'rrule';
 import dateUtils from '../../core/utils/date';
 import timeZoneUtils from './utils.timeZone.js';
-import DateAdapter from './dateAdapter';
 
 const toMs = dateUtils.dateToMilliseconds;
 
@@ -39,10 +38,10 @@ class RecurrenceProcessor {
             return result;
         }
 
-        const startDateUtc = this._getRRuleUtcDate(options.start);
-        const endDateUtc = this._getRRuleUtcDate(options.end);
-        const minDateUtc = this._getRRuleUtcDate(options.min);
-        const maxDateUtc = this._getRRuleUtcDate(options.max);
+        const startDateUtc = timeZoneUtils.createUTCDate(options.start);
+        const endDateUtc = timeZoneUtils.createUTCDate(options.end);
+        const minDateUtc = timeZoneUtils.createUTCDate(options.min);
+        const maxDateUtc = timeZoneUtils.createUTCDate(options.max);
 
         const duration = endDateUtc ? endDateUtc.getTime() - startDateUtc.getTime() : 0;
 
@@ -54,7 +53,7 @@ class RecurrenceProcessor {
             const endAppointmentTime = date.getTime() + duration;
 
             if(endAppointmentTime >= minTime) {
-                const correctDate = this._getCorrectDateByTimezoneOffset(date);
+                const correctDate = timeZoneUtils.createDateFromUTC(date);
                 result.push(correctDate);
             }
         });
@@ -183,7 +182,7 @@ class RecurrenceProcessor {
             const splitDates = options.exception.split(',');
             const exceptDates = this._getDatesByRecurrenceException(splitDates, startDateUtc);
             exceptDates.forEach(date => {
-                const utcDate = this._getRRuleUtcDate(date);
+                const utcDate = timeZoneUtils.createUTCDate(date);
                 this.rRuleSet.exdate(utcDate);
             });
         }
@@ -212,33 +211,6 @@ class RecurrenceProcessor {
         const result = [];
         ruleValues.forEach(rule => result.push(this.getDateByAsciiString(rule, date)));
         return result;
-    }
-
-    _getRRuleUtcDate(date) {
-        if(!date) {
-            return null;
-        }
-
-        const newDate = new Date(Date.UTC(
-            date.getFullYear(),
-            date.getMonth(),
-            date.getDate(),
-            date.getHours(),
-            date.getMinutes(),
-            date.getSeconds()
-        ));
-
-        return newDate;
-    }
-
-    _getCorrectDateByTimezoneOffset(date) {
-        const result = DateAdapter(date);
-
-        const timezoneOffsetBeforeInMin = result.getTimezoneOffset();
-        result.addTime(result.getTimezoneOffset('minute'));
-        result.subtractMinutes(timezoneOffsetBeforeInMin - result.getTimezoneOffset());
-
-        return result.source;
     }
 
     _parseRecurrenceRule(recurrence) {
